refactor: extract local storage store fns
