+++ plugins/browser-preferences/package.json
@@ -1,101 +1,109 @@
 {
-	"name": "@just-web/browser-preferences",
-	"version": "6.0.2",
-	"description": "browser implementation for @just-web/preferences",
-	"keywords": [
-		"just-web",
-		"framework"
-	],
-	"homepage": "https://github.com/justland/just-web/tree/main/plugins/browser-preferences",
-	"bugs": {
-		"url": "https://github.com/justland/just-web/issues"
-	},
-	"repository": {
-		"type": "git",
-		"url": "https://github.com/justland/just-web.git",
-		"directory": "plugins/browser-preferences"
-	},
-	"license": "MIT",
-	"author": {
-		"name": "Homa Wong (unional)",
-		"email": "homawong@gmail.com",
-		"url": "https://github.com/unional"
-	},
-	"sideEffects": false,
-	"type": "module",
-	"exports": {
-		"types": "./esm/index.d.ts",
-		"import": "./esm/index.js",
-		"require": "./cjs/index.js",
-		"default": "./cjs/index.js"
-	},
-	"main": "./cjs/index.js",
-	"module": "./esm/index.js",
-	"types": "./esm/index.d.ts",
-	"files": [
-		"cjs",
-		"esm",
-		"ts",
-		"!**/*.{spec,test,unit,accept,integrate,system}.*"
-	],
-	"scripts": {
-		"build": "run-p build:cjs build:esm",
-		"build:cjs": "tsc -p tsconfig.cjs.json && ncp ../../config/package.cjs.json cjs/package.json",
-		"build:esm": "tsc",
-		"build:watch": "tsc -w",
-		"clean": "rimraf cjs esm coverage lib *.tsbuildinfo",
-		"coverage": "cross-env NODE_OPTIONS=--experimental-vm-modules NODE_NO_WARNINGS=1 jest --coverage",
-		"depcheck": "depcheck",
-		"lint": "cross-env TIMING=1 eslint --ext=ts,tsx,js,yaml,yml .",
-		"nuke": "rimraf node_modules",
-		"size": "size-limit",
-		"test": "cross-env NODE_OPTIONS=--experimental-vm-modules jest",
-		"test:watch": "cross-env NODE_OPTIONS=--experimental-vm-modules NODE_NO_WARNINGS=1 jest --watch",
-		"verify": "npm-run-all -p build lint coverage depcheck -p size"
-	},
-	"dependencies": {
-		"@just-web/states": "workspace:^",
-		"immer": "npm:@unional/immer@^9.0.0-beta.2",
-		"js-base64": "^3.7.3",
-		"type-plus": "^6.0.0"
-	},
-	"devDependencies": {
-		"@just-web/app": "workspace:^",
-		"@just-web/commands": "workspace:^",
-		"@just-web/keyboard": "workspace:^",
-		"@just-web/log": "workspace:^",
-		"@just-web/preferences": "workspace:^",
-		"@just-web/repo-scripts": "workspace:^",
-		"@just-web/types": "workspace:^",
-		"@repobuddy/jest": "^3.0.0",
-		"@size-limit/preset-small-lib": "~8.2.0",
-		"@types/jest": "^29.2.3",
-		"assertron": "^11.2.0",
-		"cross-env": "^7.0.3",
-		"depcheck": "~1.4.3",
-		"eslint": "^8.11.0",
-		"eslint-plugin-harmony": "^7.1.2",
-		"jest": "^29.5.0",
-		"jest-validate": "^29.5.0",
-		"jest-watch-suspend": "^1.1.2",
-		"jest-watch-toggle-config-2": "^2.1.0",
-		"jest-watch-typeahead": "^2.2.2",
-		"ncp": "^2.0.0",
-		"npm-run-all": "^4.1.5",
-		"rimraf": "~3.0.2",
-		"satisfier": "^5.4.2",
-		"size-limit": "~8.2.0",
-		"ts-jest": "^29.1.0",
-		"typescript": "^5.0.0"
-	},
-	"peerDependencies": {
-		"@just-web/commands": "workspace:^",
-		"@just-web/log": "workspace:^",
-		"@just-web/preferences": "workspace:^",
-		"@just-web/types": "workspace:^"
-	},
-	"typedoc": {
-		"entryPoint": "./ts/index.ts",
-		"readmeFile": "./README.md"
-	}
+  "name": "@just-web/browser-preferences",
+  "version": "6.0.2",
+  "description": "browser implementation for @just-web/preferences",
+  "keywords": [
+    "just-web",
+    "framework"
+  ],
+  "homepage": "https://github.com/justland/just-web/tree/main/plugins/browser-preferences",
+  "bugs": {
+    "url": "https://github.com/justland/just-web/issues"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/justland/just-web.git",
+    "directory": "plugins/browser-preferences"
+  },
+  "license": "MIT",
+  "author": {
+    "name": "Homa Wong (unional)",
+    "email": "homawong@gmail.com",
+    "url": "https://github.com/unional"
+  },
+  "sideEffects": false,
+  "type": "module",
+  "exports": {
+    ".": {
+      "import": {
+        "types": "./esm/index.d.ts",
+        "default": "./esm/index.js"
+      },
+      "require": {
+        "types": "./cjs/index.d.ts",
+        "default": "./cjs/index.js"
+      }
+    }
+  },
+  "main": "./cjs/index.js",
+  "module": "./esm/index.js",
+  "types": "./cjs/index.d.ts",
+  "files": [
+    "cjs",
+    "esm",
+    "ts",
+    "!**/*.{spec,test,unit,accept,integrate,system}.*"
+  ],
+  "scripts": {
+    "build": "run-p build:cjs build:esm",
+    "build:cjs": "tsc -p tsconfig.cjs.json && ncp ../../config/package.cjs.json cjs/package.json",
+    "build:esm": "tsc",
+    "build:watch": "tsc -w",
+    "clean": "rimraf cjs esm coverage lib *.tsbuildinfo",
+    "coverage": "cross-env NODE_OPTIONS=--experimental-vm-modules NODE_NO_WARNINGS=1 jest --coverage",
+    "depcheck": "depcheck",
+    "lint": "cross-env TIMING=1 eslint --ext=ts,tsx,js,yaml,yml .",
+    "nuke": "rimraf node_modules",
+    "size": "size-limit",
+    "test": "cross-env NODE_OPTIONS=--experimental-vm-modules jest",
+    "test:watch": "cross-env NODE_OPTIONS=--experimental-vm-modules NODE_NO_WARNINGS=1 jest --watch",
+    "w": "pnpm test:watch",
+    "verify": "npm-run-all -p build lint coverage depcheck -p size"
+  },
+  "dependencies": {
+    "@just-web/states": "workspace:^",
+    "immer": "npm:@unional/immer@^9.0.0-beta.2",
+    "js-base64": "^3.7.3",
+    "type-plus": "^6.0.0"
+  },
+  "devDependencies": {
+    "@just-web/app": "workspace:^",
+    "@just-web/commands": "workspace:^",
+    "@just-web/keyboard": "workspace:^",
+    "@just-web/log": "workspace:^",
+    "@just-web/preferences": "workspace:^",
+    "@just-web/repo-scripts": "workspace:^",
+    "@just-web/types": "workspace:^",
+    "@repobuddy/jest": "^3.0.0",
+    "@size-limit/preset-small-lib": "~8.2.0",
+    "@types/jest": "^29.2.3",
+    "assertron": "^11.2.0",
+    "cross-env": "^7.0.3",
+    "depcheck": "~1.4.3",
+    "eslint": "^8.11.0",
+    "eslint-plugin-harmony": "^7.1.2",
+    "jest": "^29.5.0",
+    "jest-validate": "^29.5.0",
+    "jest-watch-suspend": "^1.1.2",
+    "jest-watch-toggle-config-2": "^2.1.0",
+    "jest-watch-typeahead": "^2.2.2",
+    "ncp": "^2.0.0",
+    "npm-run-all": "^4.1.5",
+    "rimraf": "~3.0.2",
+    "satisfier": "^5.4.2",
+    "size-limit": "~8.2.0",
+    "ts-jest": "^29.1.0",
+    "typescript": "^5.0.0"
+  },
+  "peerDependencies": {
+    "@just-web/app": "workspace:^",
+    "@just-web/commands": "workspace:^",
+    "@just-web/log": "workspace:^",
+    "@just-web/preferences": "workspace:^",
+    "@just-web/types": "workspace:^"
+  },
+  "typedoc": {
+    "entryPoint": "./ts/index.ts",
+    "readmeFile": "./README.md"
+  }
 }
